Move partialbuffer up to utils
